# contracts/openapi/guide-tour.yaml
openapi: 3.0.3
info:
  title: LocalGuide Guide and Tour API
  version: 0.1.0
servers:
  - url: http://localhost:8080/api/v1
paths:
  /guides/{id}:
    get:
      summary: Get guide profile with tours
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuideProfileResponse'
  /guides/me:
    get:
      summary: Get current guide profile with tours
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuideProfileResponse'
    put:
      summary: Update current guide profile
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGuideProfileRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuideSummaryResponse'
  /guides/me/photo:
    post:
      summary: Upload current guide profile photo
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [ file ]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuidePhotoUploadResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UpdateGuideProfileRequest:
      type: object
      properties:
        city: { type: string }
        language: { type: string }
        bio: { type: string }
        photoUrl: { type: string }
    GuideSummaryResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        firstName: { type: string }
        lastName: { type: string }
        city: { type: string }
        language: { type: string }
        bio: { type: string }
        photoUrl: { type: string }
    GuidePhotoUploadResponse:
      type: object
      properties:
        photoUrl: { type: string }
    GuideProfileResponse:
      allOf:
        - $ref: '#/components/schemas/GuideSummaryResponse'
        - type: object
          properties:
            tours:
              type: array
              items:
                $ref: '#/components/schemas/TourResponse'
...
    TourResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        guideId: { type: integer, format: int64 }
        guideName: { type: string }
        title: { type: string }
        description: { type: string }
        category: { type: string }
        city: { type: string }
        language: { type: string }
        price: { type: number, format: decimal }
        durationHours: { type: number, format: decimal }
        groupSize: { type: integer }
...